fix: harden balance engine contract
