+++ .github/workflows/balance-engine-unit-tests.yml
@@ -1,0 +1,47 @@
+name: Balance Engine Unit Tests
+
+on:
+  pull_request:
+    paths:
+      - "packages/balance-engine/**"
+      - "package.json"
+      - "bun.lock"
+      - ".github/workflows/balance-engine-unit-tests.yml"
+  push:
+    branches: [dev]
+    paths:
+      - "packages/balance-engine/**"
+      - "package.json"
+      - "bun.lock"
+      - ".github/workflows/balance-engine-unit-tests.yml"
+
+permissions:
+  contents: read
+
+concurrency:
+  group: balance-engine-unit-tests-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  unit-tests:
+    name: Unit Tests
+    runs-on: blacksmith-2vcpu-ubuntu-2404
+    timeout-minutes: 5
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Bun
+        uses: oven-sh/setup-bun@v2
+        with:
+          bun-version: 1.3.14
+
+      - name: Install dependencies
+        run: bun install --frozen-lockfile
+
+      - name: Run unit tests
+        run: bun -F @autumn/balance-engine test:unit
+
+      - name: Run TypeScript type check
+        run: bun -F @autumn/balance-engine ts
